- Ongoing IdGenerator refactoring

+++ ums-api/src/main/java/org/ums/domain/model/immutable/LoggerEntry.java
@@ -7,8 +7,7 @@ import org.ums.domain.model.mutable.MutableLoggerEntry;
 import java.io.Serializable;
 import java.util.Date;
 
-public interface LoggerEntry extends Serializable, Identifier<Integer>,
-    EditType<MutableLoggerEntry> {
+public interface LoggerEntry extends Serializable, Identifier<Long>, EditType<MutableLoggerEntry> {
   String getSql();
 
   String getUserName();
