+++ ums-api/src/main/java/org/ums/manager/NavigationManager.java
@@ -6,7 +6,7 @@ import org.ums.domain.model.immutable.Navigation;
 import java.util.List;
 import java.util.Set;
 
-public interface NavigationManager extends ContentManager<Navigation, MutableNavigation, Integer> {
+public interface NavigationManager extends ContentManager<Navigation, MutableNavigation, Long> {
   List<Navigation> getByPermissions(final Set<String> pPermissions);
 
   List<Navigation> getByPermissionsId(final Set<Integer> pPermissionIds);
